Update mini-map and ruler; optimizations
The mini-map is now the bird-view for the scene, and will show the player's position and the selected radar images.
Fix the appearance of the tick marks on ruler, and delete all the empty update functions found.
Need to fix the mini-map teleport function later, and need to add annotation function on mini-map.